# api-gateway/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Digital Asset Marketplace API
  description: |
    REST API documentation for the **Digital Asset Marketplace** — a microservices-based platform
    for buying and selling AI-generated digital assets.

    ## Architecture
    The system consists of three microservices orchestrated through an API Gateway (nginx):
    - **Identity Service** — User registration, authentication, and profile management
    - **Catalog Service** — Digital asset listing, search, and management
    - **Order Service** — Order creation, payment processing, and order tracking

    ## Authentication
    Most endpoints require a JWT Bearer token obtained from the login endpoint.
    Include the token in the `Authorization` header: `Bearer <your_token>`
  version: 1.0.0
  contact:
    name: Digital Asset Marketplace Team
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local Development (API Gateway)
  - url: http://localhost:3001
    description: Identity Service (Direct)
  - url: http://localhost:3002
    description: Catalog Service (Direct)
  - url: http://localhost:3003
    description: Order Service (Direct)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from the Login endpoint
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "error message"

    # ── Identity Service ──
    RegisterRequest:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          minLength: 6
          example: "securepassword"
        role:
          type: string
          enum: [BUYER, CREATOR]
          example: "BUYER"
          description: "Only BUYER or CREATOR allowed. ADMIN cannot self-register."
    RegisterResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        email:
          type: string
          format: email
          example: "user@example.com"
        role:
          type: string
          example: "BUYER"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          example: "securepassword"
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT token to use for authenticated requests
          example: "eyJhbGciOiJIUzI1NiIs..."
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        email:
          type: string
          format: email
          example: "user@example.com"
        role:
          type: string
          enum: [BUYER, CREATOR, ADMIN]
          example: "CREATOR"
        created_at:
          type: string
          format: date-time
          example: "2026-06-06T09:24:00Z"

    # ── Catalog Service ──
...
paths:
...
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      description: |
        Creates a new user account. Only **BUYER** and **CREATOR** roles are allowed.
        ADMIN accounts cannot be self-registered.
        Password must be at least 6 characters.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Invalid input (bad email, weak password, or invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalid_role:
                  value: { "error": "Invalid role. Must be BUYER or CREATOR" }
                weak_password:
                  value: { "error": "Password must be at least 6 characters" }
                invalid_email:
                  value: { "error": "Invalid email format" }
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Login and get JWT token
      description: Authenticates a user and returns a JWT token for subsequent requests.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/me:
    get:
      tags: [Authentication]
      summary: Get current user profile
      description: Returns the profile of the currently authenticated user.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Catalog / Assets ──
...
  /api/admin/users:
    get:
      tags: [Admin]
      summary: List all users
      description: Returns a list of all users. Requires ADMIN role.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of users
        '403':
          description: Forbidden (Not an admin)
  /api/admin/users/{id}/suspend:
    put:
      tags: [Admin]
      summary: Suspend user account
      description: Suspends a user account so they can no longer login. Requires ADMIN role.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User suspended successfully
        '403':
          description: Forbidden (Not an admin)
        '404':
          description: User not found
  /api/admin/users/{id}/unsuspend:
    put:
      tags: [Admin]
      summary: Reactivate user account
      description: Reactivates a previously suspended user account. Requires ADMIN role.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User reactivated successfully
        '403':
          description: Forbidden (Not an admin)
        '404':
          description: User not found
...
  /api/auth/password:
    put:
      tags: [Authentication]
      summary: Change Password
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
                  minLength: 6
      responses:
        '200':
          description: Password changed
        '400':
          description: Bad request or incorrect current password
  /api/auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Forgot Password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Reset link sent
  /api/auth/reset-password:
    post:
      tags: [Authentication]
      summary: Reset Password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token:
                  type: string
                new_password:
                  type: string
                  minLength: 6
      responses:
        '200':
          description: Password reset successful
        '400':
          description: Invalid or expired token
...